refactor: move to AdminUseTableCard
