+++ apps/dashboard/src/modules/user/components/UserSettingsComponent.vue
@@ -55,7 +55,7 @@
 </template>
 
 <script setup lang="ts">
-import { ref, defineProps } from 'vue';
+import { ref } from 'vue';
 import { useI18n } from 'vue-i18n';
 import type { UserResponse } from '@sudosos/sudosos-client';
 import Divider from 'primevue/divider';
